All compatible fonts are now selected through a combobox
